Fixed Bug 10414: "Thin multisegmented custom lines incorrectly rendered to PDF"
git-svn-id: svn://scribus.net/trunk/Scribus@17048 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pdflib_core.cpp
@@ -1995,13 +1995,13 @@ bool PDFLibCore::PDF_TemplatePage(const ScPage* pag, bool )
 					QVector<double>::iterator it;
 					for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 					{
-						int da = static_cast<int>(*it);
+						double da = *it;
 						// #8758: Custom dotted lines don't export properly to pdf
 						// Null values have to be exported if line end != flat
 						if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 							PutPage(QString::number(da)+" ");
 					}
-					PutPage("] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n");
+					PutPage("] "+QString::number(ite->DashOffset)+" d\n");
 				}
 				else
 					PutPage("["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n");
@@ -3490,13 +3490,13 @@ QString PDFLibCore::PDF_ProcessTableItem(PageItem* ite, const ScPage* pag)
 		QVector<double>::iterator it;
 		for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 		{
-			int da = static_cast<int>(*it);
+			double da = *it;
 			// #8758: Custom dotted lines don't export properly to pdf
 			// Null values have to be exported if line end != flat
 			if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 				tmp += QString::number(da)+" ";
 		}
-		tmp += "] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n";
+		tmp += "] "+QString::number(ite->DashOffset)+" d\n";
 	}
 	else
 		tmp += "["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n";
@@ -3630,13 +3630,13 @@ bool PDFLibCore::PDF_ProcessItem(QString& output, PageItem* ite, const ScPage* p
 		QVector<double>::iterator it;
 		for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 		{
-			int da = static_cast<int>(*it);
+			double da = *it;
 			// #8758: Custom dotted lines don't export properly to pdf
 			// Null values have to be exported if line end != flat
 			if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 				tmp += QString::number(da)+" ";
 		}
-		tmp += "] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n";
+		tmp += "] "+QString::number(ite->DashOffset)+" d\n";
 	}
 	else
 		tmp += "["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n";
@@ -5106,13 +5106,13 @@ bool PDFLibCore::setTextCh(PageItem *ite, uint PNr, double x,  double y, uint d,
 				QVector<double>::iterator it;
 				for ( it = embedded->DashValues.begin(); it != embedded->DashValues.end(); ++it )
 				{
-					int da = static_cast<int>(*it);
+					double da = *it;
 					// #8758: Custom dotted lines don't export properly to pdf
 					// Null values have to be exported if line end != flat
 					if ((da != 0) || (embedded->lineEnd() != Qt::FlatCap))
 						tmp2 += QString::number(da)+" ";
 				}
-				tmp2 += "] "+QString::number(static_cast<int>(embedded->DashOffset))+" d\n";
+				tmp2 += "] "+QString::number(embedded->DashOffset)+" d\n";
 			}
 			else
 				tmp2 += "["+getDashString(embedded->PLineArt, embedded->lineWidth())+"] 0 d\n";
@@ -6223,13 +6223,13 @@ bool PDFLibCore::PDF_PatternFillStroke(QString& output, PageItem *currItem, int 
 			QVector<double>::iterator it;
 			for ( it = item->DashValues.begin(); it != item->DashValues.end(); ++it )
 			{
-				int da = static_cast<int>(*it);
+				double da = *it;
 				// #8758: Custom dotted lines don't export properly to pdf
 				// Null values have to be exported if line end != flat
 				if ((da != 0) || (item->lineEnd() != Qt::FlatCap))
 					tmp2 += QString::number(da)+" ";
 			}
-			tmp2 += "] "+QString::number(static_cast<int>(item->DashOffset))+" d\n";
+			tmp2 += "] "+QString::number(item->DashOffset)+" d\n";
 		}
 		else
 			tmp2 += "["+getDashString(item->PLineArt, item->lineWidth())+"] 0 d\n";
